Unmarshaling objects: ALLOCATE + INTIALIZE
Consider this class:

    (defclass bad ()
      ((mandatory
          :initform (error "Argument must be provided")
          :initarg :mandatory)
       (something :initform (random))))

Assume also that MANDATORY is a persistent slot and SOMETHING isn't:

    (defmethod marshal:class-persistant-slots ((o bad))
      '(mandatory))

UNMARHSAL signals an error because it uses MAKE-INSTANCE. We need to use
ALLOCATE-INSTANCE instead. However, in that case SOMETHING would not get
properly initialized. That's why the commit also changes the default
method of INITIALIZE-UNMARSHALLED-INSTANCE, so that the object's slots
which are not set by UNMARSHAL can take default values.