- signalling a more useful error  message when marshalling an instance   of a class with no serializable slots indicated:
When a  class with no method  ' ms:class-persistent-slots' specialized
is marshalled the serialized structure  fail to be restored signalling
an error not really useful to understand what is going on.

Example code:

-------------------------------------
(in-package :cl-user)

(require :marshal)

(use-package :marshal)

(defclass a () ())

(defclass b ()
  ((some-a
    :accessor some-a
    :initarg :some-a
    :initform nil)))

(defmethod ms:class-persistent-slots ((object b))
  '(some-a))

(defun test-bug ()
  (let ((instance (make-instance 'b
                                 :some-a (make-instance 'a))))
    (format t "~s" (ms:marshal instance))
    (ms:unmarshal (ms:marshal instance))))

(test-bug) ; =>

; The index 0 is too large for a list of length 0.
;   [Condition of type SB-KERNEL:INDEX-TOO-LARGE-ERROR]

--------------------

With this patch a better (i hope so!) error message is signalled.

Thanks to people from IRC for reporting this!